Introduced input flag to CLI.
Current implementation replaces prompt function with a mutating
function based on the input string if input flag activated.

This feels like a flawed/impure approach but it actually works
as an implementation.

If integrated into src/index.js, it would be within the scope
of automated testing as it currently stands, but wanted to test
that the idea would work first.