Hid prompts so only output shows when input already provided.
This is the simplest way to allow results of crafting diffs
can be piped directly to another command (key example being
to pipe patch output to git apply --cached). This is achieved
simply by disabling logging and printing only the result patch.

However, there is bound to be a better way involving changing
the way outputs are processed.