Separated logging output from result output
Used the same principle that was at work in the use of prompts
-- namely used the ttys package to specify a stdout route as well
as stdin.

So interactivity can still display in the terminal but it will
not be part of the output if piping is used after the command is
run. Instead only the resultant patch will be piped to the next
command via console.log. (And errors should be handled via
console.error now too.)