Showed interaction information again even when input predetermined.
The prompts no longer show but the read-out of hunks and diffs
being crafted do as it would in interaction. This was made
possible by the difference between logging output and piping
output. Now behaviour of -i flag is aligned with not having -i
flag, with no detriment to piping/applying patches.

(A quiet mode may be necessary at some point anyway.)